add logging to stenographer process
Adds the '-v 1' option when launching stenographer to enable verbose level 1 on stenographer.  Stenotype logging was already enabled.

Also moves the stdio redirect to the top of the script so that stenokeys.sh output also goes to logfile instead of to docker logs output.

Also adds a couple of strategic exec commands to avoid intermediate shells.